requirements.txt
discord.py
numpy
pillow
opencv-python
mss
sounddevice
textblob
google-generativeai
wave
asyncio
socketio
pyinstaller
gitpython
clipboard